backend/listener_service/requirements.txt
fastapi
uvicorn
pydantic
pyaudio
numpy
psycopg2-binary
sqlalchemy
whisper
pyannote.audio
torch
torchaudio
transformers
PyYAML